pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "tars-cli"
version = "3.1.0"
description = "AI-powered Kubernetes cluster monitoring and troubleshooting tool for SREs"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Omer Rathore", email = "orathore93@gmail.com"}
]
keywords = ["kubernetes", "monitoring", "ai", "devops", "k8s", "cli", "sre", "oncall"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "kubernetes>=28.0.0",
    "google-genai>=0.2.0",
    "prometheus-api-client>=0.5.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/orathore93-hue/tars-cli"
Repository = "https://github.com/orathore93-hue/tars-cli"
Issues = "https://github.com/orathore93-hue/tars-cli/issues"

[project.scripts]
tars = "tars.cli:main"

[tool.setuptools]
packages = ["tars"]
package-dir = {"" = "src"}
